pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "python-semantic-release>=9.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cis-bench"
version = "1.0.0"
description = "CLI tool for fetching and managing CIS benchmarks from CIS WorkBench"
authors = [
    {name = "MITRE SAF Team", email = "saf@mitre.org"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.8"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

keywords = ["cis", "benchmark", "security", "compliance", "xccdf", "scap", "nist", "mitre"]

dependencies = [
    # Core scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "urllib3>=2.0.6",
    # Data modeling
    "pydantic>=2.0.0",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.0",
    # CLI
    "click>=8.1.0",
    "rich>=13.7.0",
    "questionary>=2.0.0",
    # Authentication
    "browser-cookie3>=0.19.0",
    # SSL/TLS - Use system certificate stores
    "truststore>=0.8.0",
    # Configuration
    "python-dotenv>=1.0.0",
    # Export formats
    "PyYAML>=6.0.0",
    "xsdata[cli,lxml]>=24.0.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    # Linting and formatting
    "ruff>=0.8.0",
    # Security
    "bandit[toml]>=1.7.0",
    # Type checking (optional)
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    # Pre-commit
    "pre-commit>=4.0.0",
    # Build tools
    "build>=1.0.0",
    "twine>=5.0.0",
    # Documentation
    "mkdocs-material>=9.5.0",
]

[project.urls]
Homepage = "https://github.com/mitre/cis-bench"
Documentation = "https://mitre.github.io/cis-bench"
Repository = "https://github.com/mitre/cis-bench"
Issues = "https://github.com/mitre/cis-bench/issues"

[project.scripts]
cis-bench = "cis_bench.cli.app:cli"

[tool.setuptools]
packages = {find = {where = ["src"]}}

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
upload_to_vcs_release = true
upload_to_pypi = false  # Handled by separate workflow step
build_command = "pip install build && python -m build"
commit_message = "chore: release {version}"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "docs", "chore", "refactor", "test", "ci"]
minor_tags = ["feat"]
patch_tags = ["fix", "docs", "chore", "refactor", "test", "ci"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

# Markers for organizing tests
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (moderate speed, some I/O)",
    "e2e: End-to-end tests (slow, full pipeline)",
    "architecture: Architecture compliance tests",
]

# Output options
addopts = [
    "--strict-markers",
    "--tb=short",
    "--verbose",
]

# Coverage options
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | \.eggs
  | *.egg-info
)/
'''

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "S101",  # Allow assert in tests
    "S320",  # lxml XML parsing (we parse XML we generate ourselves - acceptable risk)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S603", "S607", "B023", "F841"]  # Allow test-specific patterns
"src/cis_bench/models/xccdf/*.py" = ["S105"]  # PASS/FAIL are test result enums, not passwords
"src/cis_bench/models/xccdf_v1_1/*.py" = ["S105"]  # PASS/FAIL are test result enums, not passwords
"fetch.py" = ["B023"]  # Legacy script - not fixing
"src/cis_bench/cli/commands/download.py" = ["B023"]  # Loop variable in callback is intentional
"src/cis_bench/cli/app.py" = ["E402"]  # Imports after function to avoid circular imports with Click

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradually enable

[tool.bandit]
# Bandit security scanner configuration
exclude_dirs = ["/tests", "/venv", "/.venv", "/build", "/dist"]
skips = [
    "B101",  # Skip assert_used (allowed in tests)
    "B105",  # Skip hardcoded_password_string (false positives in XCCDF enums: PASS/FAIL)
    "B320",  # Skip lxml.etree.fromstring warnings (we parse XML we generate ourselves)
    "B404",  # Skip subprocess import (used in tests for xmllint validation)
    "B410",  # Skip lxml import warnings (required for XCCDF processing)
    "B602",  # Skip shell=True (needed for Windows 'start' command - URL is trusted)
    "B603",  # Skip subprocess without shell (we use hardcoded commands)
    "B607",  # Skip partial executable path (xmllint is in PATH)
]
